refactor: create classification agent file
